fix: don't warn for high memory usage when using the network
