+++ crates/core/machine/src/utils/logger.rs
@@ -1,10 +1,5 @@
-use std::{
-    sync::Once,
-    thread::{sleep, spawn},
-    time::Duration,
-};
+use std::sync::Once;
 
-use sysinfo::{MemoryRefreshKind, RefreshKind, System};
 use tracing_forest::ForestLayer;
 use tracing_subscriber::{
     fmt::format::FmtSpan, layer::SubscriberExt, util::SubscriberInitExt, EnvFilter, Registry,
@@ -49,29 +44,5 @@ pub fn setup_logger() {
                 panic!("Invalid logger type: {logger_type}");
             }
         };
-
-        // Spawn a thread to warn when used memory is high
-        spawn(|| {
-            let mut sys = System::new_with_specifics(
-                RefreshKind::new().with_memory(MemoryRefreshKind::new().with_ram()),
-            );
-
-            let total_memory = sys.total_memory();
-
-            loop {
-                sleep(Duration::from_secs(10));
-                sys.refresh_memory();
-
-                let used_memory = sys.used_memory();
-                let ratio = used_memory as f64 / total_memory as f64;
-
-                if ratio > 0.8 {
-                    tracing::warn!(
-                        "Memory usage is high: {:.2}%, we recommand using the prover network",
-                        ratio * 100.0
-                    );
-                }
-            }
-        });
     });
 }
