+++ crates/sdk/Cargo.toml
@@ -51,6 +51,7 @@ tonic = { version = "0.12", features = ["tls", "tls-roots"], optional = true }
 alloy-sol-types = { version = "1.0", default-features = false }
 alloy-primitives = { version = "1.0", default-features = false, optional = true, features = ["k256", "serde"] }
 backoff = { version = "0.4", features = ["tokio"], optional = true }
+sysinfo = "0.30.13"
 
 # TEE Dependencies
 k256 = { version = "0.13.3", features = ["serde"] } # Signing
